Fold the account lookup into /team user
One command, not two: a staffer looking someone up should not have to know
which of two commands holds the field they need. /team user now renders the
Discord identity, the Moddy row, the subscription, the staff roles, the
attributes, the global sanction level, the case counts and the notifications.

The gate moves from the command to the section: /team user stays open to all
staff as it always was, and email, Stripe customer and stored preferences
render only for a caller holding the user_lookup node, re-derived per
invocation and failing closed. Without it the section names the missing node.
