+++ CLAUDE.md
@@ -140,8 +140,8 @@ moddy/
 │   │   ├── design.py          #     Standardized Components V2 panels (accents)
 │   │   └── parsing.py         #     Arg helpers (user/guild id)
 │   ├── commands/dev/          #   /dev commands (one file each)
-│   ├── commands/team/         #   /team commands (incl. help, /team account —
-│   │                          #   a user's Moddy account: email, billing, cases)
+│   ├── commands/team/         #   /team commands (incl. help, /team user — full
+│   │                          #   user card: account, billing, cases, notifications)
 │   ├── commands/mod/          #   /mod commands + case/, global/ and altguard/ sub-groups
 │   │                          #   (incl. /mod notif — notification lookup by uuid)
 │   ├── commands/com/          #   /com commands (/com send — notification to a user,
@@ -275,7 +275,7 @@ moddy/
     ├── test_logs.py           #   Server logs: registry, routing, rendering, delivery
     ├── test_logs_i18n.py      #   Server logs: i18n completeness on the 5 locales
     ├── test_heartbeat.py      #   Health Monitor heartbeat: payload, lifecycle, status decisions
-    └── test_staff_account.py  #   /team account: sections, node gating, i18n on the 5 locales
+    └── test_staff_user_command.py # /team user: sections, personal-data gate, i18n (5 locales)
 ```
 
 ---
