+++ docs/STAFF_SYSTEM.md
@@ -100,7 +100,7 @@ no nodes has no command access even if it matches the type check.
 | `redirect_manage` | `/manage redirect *` |
 | `banner_manage` | `/manage banner *` |
 | `official_manage` | `/dev official` |
-| `user_lookup` | `/team account` — a user's Moddy account: email, Stripe customer, subscription, staff roles, cases, notifications. Personal data, so it is gated even though `/team` is open to all staff |
+| `user_lookup` | The personal-data section of `/team user` (email, Stripe customer, stored preferences). The command itself stays open to all staff — only that section is gated |
 | `support_request` | Claim / reply to / close a `/bug-report` or configuration-help request ([SUPPORT_REQUESTS.md](SUPPORT_REQUESTS.md)) |
 | `broadcast` | `/com send`, `/com beta` |
