+++ staff/commands/team/user.py
@@ -1,18 +1,76 @@
-"""`/team user` — detailed information about a user (Discord + Moddy data)."""
+"""`/team user` — everything Moddy knows about a user.
+
+One card for the whole staff surface: the Discord account, the Moddy row
+behind it (email, Stripe customer, first seen), the subscription, the staff
+roles, the attributes, the global sanction level, the cases opened against
+them and the notifications Moddy sent them.
+
+The email, the Stripe customer and the stored preferences are personal data,
+so that one section is only rendered for staffers holding the ``user_lookup``
+node (see ``utils/staff_role_permissions.py``). Everything else stays
+available to every staff member, as it has always been, and the invocation
+goes through the usual staff audit log either way.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import logging
+from datetime import datetime, timezone
+from typing import Any, Dict, List, Optional
 
 import discord
 
-from staff.framework import StaffCommand, SlashOption, staff_command, design, CommandType, parse_user_id
+from staff.framework import (
+    StaffCommand, SlashOption, staff_command, design, CommandType, parse_user_id,
+)
 from staff.framework import badges
-from utils import emojis
+from utils import emojis, global_sanctions
 from utils.i18n import t
+from utils.staff_permissions import has_staff_node
+
+logger = logging.getLogger("moddy.staff.team.user")
+
+#: How many recent notifications the panel summarizes.
+_NOTIFICATION_SAMPLE = 25
+
+#: Node unlocking the personal-data section (email, billing, preferences).
+SENSITIVE_NODE = "user_lookup"
+
+
+def _stamp(moment: Optional[datetime], style: str = "R") -> str:
+    """``<t:…>`` for a datetime, ``—`` when there is none."""
+    if not moment:
+        return "`—`"
+    if moment.tzinfo is None:
+        moment = moment.replace(tzinfo=timezone.utc)
+    return f"<t:{int(moment.timestamp())}:{style}>"
+
+
+def _code(value: Any) -> str:
+    """Backtick a dynamic value (DESIGN.md rule), ``—`` when empty."""
+    text = str(value) if value not in (None, "") else "—"
+    return f"`{text}`"
+
+
+async def _safe(coro, default):
+    """Await a DB call, degrading to ``default`` instead of failing the panel.
+
+    One unavailable table must not cost the staffer the eight other sections.
+    """
+    try:
+        return await coro
+    except Exception as exc:  # pragma: no cover - defensive
+        logger.warning("user lookup: %s", exc, exc_info=True)
+        return default
 
 
 @staff_command
 class UserCommand(StaffCommand):
     command_type = CommandType.TEAM
     name = "user"
-    description = "Detailed info about a user (Discord + Moddy data)."
+    aliases = ("whois",)
+    description = "Everything Moddy knows about a user (account, billing, cases…)."
     options = [
         SlashOption("user", "user", "Target user.", required=True),
     ]
@@ -27,8 +85,10 @@ class UserCommand(StaffCommand):
         target = ctx.opt("user")
         user_id = target.id if target else parse_user_id(ctx.opt("user_id") or "")
         if not user_id:
-            await ctx.send(view=design.invalid_usage(locale, "t.user <user_id>"))
+            await ctx.send(view=design.invalid_usage(locale, "t.user <@user|user_id>"))
             return
+
+        await ctx.defer()
 
         try:
             user = await bot.fetch_user(user_id)
@@ -39,49 +99,187 @@ class UserCommand(StaffCommand):
             ))
             return
 
-        attributes, verification = await badges.fetch_verification(bot, user_id)
+        if not bot.db:
+            await ctx.send(view=design.error(
+                t("staff.common.error.title", locale=locale),
+                t("staff.team.user.no_db", locale=locale),
+            ))
+            return
+
+        from utils.subscription import get_subscription
+
+        (record, staff_perms, subscription, sub_servers, cases_total, cases_open,
+         scopes, global_actions, notifications, may_see_personal) = await asyncio.gather(
+            _safe(bot.db.get_user(user_id), {}),
+            _safe(bot.db.get_staff_permissions(user_id), {}),
+            _safe(get_subscription(bot, user_id), None),
+            _safe(bot.db.get_subscription_servers(user_id), []),
+            _safe(bot.db.count_subject_cases("discord_user", user_id), 0),
+            _safe(bot.db.count_subject_cases("discord_user", user_id, "open"), 0),
+            _safe(bot.db.list_subject_scopes("discord_user", user_id), []),
+            _safe(bot.db.list_active_global_actions("discord_user", user_id), []),
+            _safe(bot.db.list_notifications(recipient_id=user_id, limit=_NOTIFICATION_SAMPLE), []),
+            _safe(has_staff_node(bot, ctx.author.id, SENSITIVE_NODE), False),
+        )
+
+        attributes = (record or {}).get("attributes") or {}
+        verification = ((record or {}).get("data") or {}).get("verification") or {}
         rendered, orgs, tier = badges.render_name(user, attributes, verification)
-
-        fields = [{
-            "name": f"{emojis.USER} {t('staff.team.user.basic', locale=locale)}",
-            "value": (
-                f"**ID:** `{user.id}`\n"
-                f"**{t('staff.team.user.username', locale=locale)}:** `@{user.name}`\n"
-                f"**{t('staff.team.user.bot', locale=locale)}:** `{'yes' if user.bot else 'no'}`\n"
-                f"**{t('staff.team.user.created', locale=locale)}:** <t:{int(user.created_at.timestamp())}:R>"
-            ),
-        }]
-
-        if attributes:
-            attr_lines = [f"`{k}`" + (f": `{v}`" if v is not True else "") for k, v in attributes.items()]
-            fields.append({"name": f"{emojis.SETTINGS} {t('staff.team.attributes', locale=locale)}",
-                           "value": " • ".join(attr_lines)})
-        else:
-            fields.append({"name": f"{emojis.SETTINGS} {t('staff.team.attributes', locale=locale)}",
-                           "value": f"-# {t('staff.team.none', locale=locale)}"})
-
-        shared = sum(1 for g in bot.guilds if g.get_member(user_id))
-        fields.append({"name": f"{emojis.WEB} {t('staff.team.user.shared', locale=locale)}",
-                       "value": f"`{shared}`"})
-
-        if bot.db:
-            try:
-                db_data = await bot.db.get_user(user_id)
-                if db_data and db_data.get("created_at"):
-                    fields.append({"name": f"{emojis.TIME} {t('staff.team.user.first_seen', locale=locale)}",
-                                   "value": f"<t:{int(db_data['created_at'].timestamp())}:R>"})
-            except Exception:
-                pass
 
         description = rendered
         if tier == "org_member" and orgs:
             description += f"\n-# {t('staff.common.affiliation', locale=locale, orgs=', '.join(orgs))}"
+
+        fields = [
+            _identity_field(bot, user, locale),
+            _account_field(record, may_see_personal, locale),
+            _subscription_field(subscription, sub_servers, locale),
+            _attributes_field(attributes, locale),
+            _moderation_field(global_actions, cases_total, cases_open, scopes, locale),
+            _notifications_field(notifications, locale),
+        ]
+        staff_field = _staff_field(staff_perms, locale)
+        if staff_field:
+            fields.insert(3, staff_field)
 
         await ctx.send(view=design.panel(
             "info",
             t("staff.team.user.title", locale=locale),
             description,
             fields=fields,
-            emoji=emojis.USER,
+            emoji=emojis.MANAGE_USER,
             accent="primary",
+            footer=t("staff.team.user.privacy", locale=locale) if may_see_personal else None,
         ))
+
+
+# --------------------------------------------------------------------------- #
+# Sections
+# --------------------------------------------------------------------------- #
+
+def _identity_field(bot, user: discord.abc.User, locale: str) -> Dict[str, str]:
+    shared = sum(1 for guild in bot.guilds if guild.get_member(user.id))
+    return {
+        "name": f"{emojis.USER} {t('staff.team.user.basic', locale=locale)}",
+        "value": (
+            f"**ID:** `{user.id}`\n"
+            f"**{t('staff.team.user.username', locale=locale)}:** `@{user.name}`\n"
+            f"**{t('staff.team.user.bot', locale=locale)}:** `{'yes' if user.bot else 'no'}`\n"
+            f"**{t('staff.team.user.created', locale=locale)}:** {_stamp(user.created_at)}\n"
+            f"**{t('staff.team.user.shared', locale=locale)}:** `{shared}`"
+        ),
+    }
+
+
+def _account_field(record: Dict[str, Any], may_see_personal: bool, locale: str) -> Dict[str, str]:
+    """The Moddy row. Email / billing / preferences need the node."""
+    name = f"{emojis.AT} {t('staff.team.user.moddy_account', locale=locale)}"
+    lines = [
+        f"**{t('staff.team.user.first_seen', locale=locale)}:** {_stamp((record or {}).get('created_at'))}",
+        f"**{t('staff.team.user.updated', locale=locale)}:** {_stamp((record or {}).get('updated_at'))}",
+    ]
+    if not may_see_personal:
+        lines.append(f"-# {t('staff.team.user.restricted', locale=locale)}")
+        return {"name": name, "value": "\n".join(lines)}
+
+    data = (record or {}).get("data") or {}
+    email = (record or {}).get("email")
+    lines.insert(0, f"**{t('staff.team.user.email', locale=locale)}:** {_code(email)}")
+    lines.insert(1, f"**Stripe:** {_code((record or {}).get('stripe_customer_id'))}")
+    if data.get("reminder_timezone"):
+        lines.append(f"**{t('staff.team.user.timezone', locale=locale)}:** "
+                     f"{_code(data['reminder_timezone'])}")
+    if not email:
+        lines.append(f"-# {t('staff.team.user.no_email', locale=locale)}")
+    return {"name": name, "value": "\n".join(lines)}
+
+
+def _subscription_field(subscription: Optional[Dict[str, Any]],
+                        servers: List[Dict[str, Any]], locale: str) -> Dict[str, str]:
+    active = bool(subscription and subscription.get("is_active"))
+    dot = emojis.GREEN_STATUS if active else emojis.RED_STATUS
+    status = t("staff.team.subscription.active" if active else "staff.team.subscription.inactive",
+               locale=locale)
+    lines = [f"{dot} {status}"]
+    if active:
+        lines[0] += f" — `{subscription.get('tier') or 'Moddy Max'}`"
+    if subscription and subscription.get("expires_at"):
+        lines.append(f"**{t('staff.team.subscription.expires', locale=locale)}:** "
+                     f"{_stamp(subscription['expires_at'], 'f')} ({_stamp(subscription['expires_at'])})")
+    lines.append(f"**{t('staff.team.subscription.servers', locale=locale)}:** `{len(servers)}/5`"
+                 + ("\n" + "\n".join(f"-# `{s['server_id']}` — {_stamp(s.get('added_at'), 'D')}"
+                                     for s in servers[:5]) if servers else ""))
+    return {
+        "name": f"{emojis.PREMIUM} {t('staff.team.subscription.title', locale=locale)}",
+        "value": "\n".join(lines),
+    }
+
+
+def _staff_field(staff_perms: Dict[str, Any], locale: str) -> Optional[Dict[str, str]]:
+    roles = (staff_perms or {}).get("roles") or []
+    if not roles:
+        return None
+    lines = [" ".join(f"{badges.role_badge(role)} `{role}`" for role in roles)]
+    nodes = sorted({node
+                    for granted in ((staff_perms or {}).get("role_permissions") or {}).values()
+                    for node in granted})
+    lines.append(f"**{t('staff.team.user.nodes', locale=locale)}:** "
+                 + (" • ".join(f"`{node}`" for node in nodes)
+                    if nodes else f"`{t('staff.team.none', locale=locale)}`"))
+    denied = (staff_perms or {}).get("denied_commands") or []
+    if denied:
+        lines.append(f"**{t('staff.team.user.denied', locale=locale)}:** "
+                     + " • ".join(f"`{cmd}`" for cmd in denied))
+    lines.append(f"-# {t('staff.team.user.staff_since', locale=locale)} "
+                 f"{_stamp((staff_perms or {}).get('created_at'))}")
+    return {
+        "name": f"{emojis.STAFF} {t('staff.team.user.staff', locale=locale)}",
+        "value": "\n".join(lines),
+    }
+
+
+def _attributes_field(attributes: Dict[str, Any], locale: str) -> Dict[str, str]:
+    if attributes:
+        value = " • ".join(f"`{key}`" + (f": `{val}`" if val is not True else "")
+                           for key, val in attributes.items())
+    else:
+        value = f"-# {t('staff.team.none', locale=locale)}"
+    return {"name": f"{emojis.SETTINGS} {t('staff.team.attributes', locale=locale)}",
+            "value": value}
+
+
+def _moderation_field(global_actions: List[Dict[str, Any]], total: int, open_count: int,
+                      scopes: List[Dict[str, Any]], locale: str) -> Dict[str, str]:
+    level = global_sanctions.level_from_actions(
+        action["action"] for action in (global_actions or []))
+    lines = [
+        f"{global_sanctions.level_emoji(level)} "
+        f"**{t('staff.team.user.global_level', locale=locale)}:** `{level.value}`"
+    ]
+    expiries = [action.get("expires_at") for action in (global_actions or [])
+                if action.get("expires_at")]
+    if expiries:
+        lines.append(f"-# {t('staff.team.user.global_expires', locale=locale)} "
+                     f"{_stamp(min(expiries), 'f')}")
+    lines.append(
+        f"**{t('staff.team.user.cases', locale=locale)}:** `{total}` "
+        f"({t('staff.team.user.cases_open', locale=locale)} `{open_count}`) — "
+        f"`{len(scopes or [])}` {t('staff.team.user.cases_scopes', locale=locale)}"
+    )
+    return {"name": f"{emojis.SHIELD} {t('staff.team.user.moderation', locale=locale)}",
+            "value": "\n".join(lines)}
+
+
+def _notifications_field(notifications: List[Dict[str, Any]], locale: str) -> Dict[str, str]:
+    name = f"{emojis.MESSAGE} {t('staff.team.user.notifications', locale=locale)}"
+    if not notifications:
+        return {"name": name, "value": f"-# {t('staff.team.none', locale=locale)}"}
+    shown = len(notifications)
+    suffix = "+" if shown >= _NOTIFICATION_SAMPLE else ""
+    latest = notifications[0]
+    return {"name": name, "value": "\n".join([
+        f"**{t('staff.team.user.notifications_recent', locale=locale)}:** `{shown}{suffix}`",
+        f"-# {t('staff.team.user.notifications_last', locale=locale)} "
+        f"`{latest.get('kind') or '—'}` — {_stamp(latest.get('created_at'))}",
+        f"-# `{latest.get('id')}`",
+    ])}
